refactor: ♻️ profile contents tabs and edit banner
